Moving evaluation logic to ExpressionEvaluation.kt
